IN-282 refactor: change file directory
